requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pytest==7.4.3
pytest-mock==3.12.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
pyjwt==2.8.0
joblib==1.3.2
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7
passlib==1.7.4
bcrypt==4.0.1
